combine test and dont use argv
